dispatch 2 Home.jsx sections in 2 components : Keywords.jsx and TinderCard.jsx
